refactor: move Direct activation from Profiles to the Apps pane
Direct activation binds a global hotkey to focus/launch a specific app — it
bypasses the switcher, so it isn't a switcher profile. Move its section (the
nine choose-app + recorder slots) into the Apps pane, next to the per-app rules
and pinned apps where app-targeted settings belong. The Profiles pane now holds
just the switcher-shortcut list and the in-panel keys; search + anchors updated.
