+++ BetterCmdTab/Settings/AppsSettingsViewController.swift
@@ -1,5 +1,6 @@
 import AppKit
 import BetterSettings
+import BetterShortcuts
 import Combine
 
 /// Settings tab gathering per-app configuration: the "App rules" list (hide /
@@ -32,6 +33,11 @@ final class AppsSettingsViewController: SettingsTabViewController {
     private var appsSheet: AppsPickerSheetWindowController?
     private var addSheet: AppsPickerSheetWindowController?
 
+    // Direct-activation slots: a "choose app" button + shortcut recorder per slot
+    // (a global hotkey that jumps straight to a chosen app, bypassing the switcher).
+    private var directButtons: [NSButton] = []
+    private var directSlotSheet: AppsPickerSheetWindowController?
+
     private var cancellables = Set<AnyCancellable>()
 
     override func setupContent() {
@@ -53,12 +59,81 @@ final class AppsSettingsViewController: SettingsTabViewController {
         register(searchTarget: rulesCard, itemID: SearchID.exceptions)
         rebuildRulesCard()
 
+        // Direct activation — global hotkeys that focus (and launch) a chosen app,
+        // bypassing the switcher. App-targeted, so it lives with the per-app config.
+        let direct = addSection(title: String(localized: "Direct activation"), anchor: SettingsAnchor.directActivation)
+        addRow(
+            to: direct,
+            title: String(localized: "Jump straight to an app"),
+            subtitle: String(localized: "Give a shortcut to one app — it focuses that app, opening it first if needed."),
+            searchItemID: SearchID.directActivation
+        )
+        for (index, name) in BetterShortcuts.Name.directActivate.enumerated() {
+            let recorder = BetterShortcuts.RecorderCocoa(for: name)
+            let button = NSButton(title: String(localized: "Choose…"), target: self, action: #selector(chooseDirectApp(_:)))
+            button.bezelStyle = .rounded
+            button.controlSize = .small
+            button.tag = index
+            let stack = NSStackView(views: [button, recorder])
+            stack.orientation = .horizontal
+            stack.spacing = 8
+            stack.alignment = .centerY
+            addRow(to: direct, title: String(localized: "Slot \(index + 1)"), accessory: stack)
+            directButtons.append(button)
+        }
+
         // Pinned apps — a lightweight chooser (a picker, not an editor).
         let pinned = addSection(title: String(localized: "Pinned"), anchor: SettingsAnchor.pinned)
         configureManageButton(pinnedButton, action: #selector(managePinned))
         pinnedRow = addRow(to: pinned, title: String(localized: "Pinned apps"),
                            subtitle: Self.pinnedDescription(Preferences.shared.pinnedBundleIDs.count),
                            accessory: pinnedButton, searchItemID: SearchID.pinnedApps)
+    }
+
+    // MARK: - Direct activation slots
+
+    /// Sync each slot's "choose app" button to its stored bundle ID.
+    private func refreshDirectSlots() {
+        let bindings = Preferences.shared.directActivationBindings
+        for (index, button) in directButtons.enumerated() {
+            let bundleID = bindings.indices.contains(index) ? bindings[index] : ""
+            if bundleID.isEmpty {
+                button.title = String(localized: "Choose…")
+                button.image = nil
+            } else {
+                let info = Self.appInfo(for: bundleID)
+                button.title = info.name
+                info.icon.size = NSSize(width: 16, height: 16)
+                button.image = info.icon
+                button.imagePosition = .imageLeading
+            }
+        }
+    }
+
+    @objc private func chooseDirectApp(_ sender: NSButton) {
+        let slot = sender.tag
+        guard let window = view.window, directSlotSheet == nil else { return }
+        let current = Preferences.shared.directActivationBindings
+        let selected: Set<String> = (current.indices.contains(slot) && !current[slot].isEmpty) ? [current[slot]] : []
+        let controller = AppsPickerSheetWindowController(
+            title: String(localized: "Activate App"),
+            prompt: String(localized: "Choose the app this shortcut focuses."),
+            selectedBundleIDs: selected,
+            singleSelection: true,
+            confirmTitle: String(localized: "Choose")
+        ) { selection in
+            var bindings = Preferences.shared.directActivationBindings
+            while bindings.count <= slot { bindings.append("") }
+            bindings[slot] = selection.sorted().first ?? ""
+            Preferences.shared.directActivationBindings = bindings
+        }
+        controller.onDidDismiss = { [weak self] in
+            self?.directSlotSheet = nil
+            self?.refreshDirectSlots()
+        }
+        directSlotSheet = controller
+        trackForRelease(controller)
+        controller.present(asSheetFor: window)
     }
 
     private func makeGroupHeader(title: String, description: String) -> NSView {
@@ -199,6 +274,7 @@ final class AppsSettingsViewController: SettingsTabViewController {
             exceptions = current
             rebuildRulesCard()
         }
+        refreshDirectSlots()
         updatePinnedCount()
         Preferences.shared.$pinnedBundleIDs
             .receive(on: DispatchQueue.main)
